Reshuffle file structure according to PR
